refactor : ApiControllerConfig 설정 변경

+++ fl-common/src/main/java/com/common/point/request/PaymentRequest.java
@@ -14,5 +14,5 @@ public class PaymentRequest {
 
     private final String token;
 
-    private final int amount;
+    private final double amount;
 }
