+++ fl-common/src/main/java/com/common/user/response/UserAccountResponse.java
@@ -17,7 +17,7 @@ public class UserAccountResponse {
 
     private String accountNumber;
 
-    private Integer point;
+    private double point;
 
     private List<PointChargeHistoryResponse> chargeHistories;
 }
